test: eliminate the filler-named test files
