+++ internal/cmds/cds/attest_workloadclaims_test.go
@@ -10,6 +10,7 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"encoding/pem"
+	"errors"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -309,4 +310,78 @@ func mustCSR(t *testing.T) string {
 	t.Helper()
 	csr, _ := generateCSR(t)
 	return csr
+}
+
+func TestAttest_RejectsMalformedWorkloadClaimsEncoding(t *testing.T) {
+	mock := newMockAttestationApi(t, "x")
+	h := newTestAttestHandler(t, mock.URL, nil)
+	csrPEM, _ := generateCSR(t)
+
+	body, err := json.Marshal(types.AttestRequestBody{
+		Challenge:      issueChallenge(t, h),
+		Evidence:       types.AttestationEvidence{Platform: "snp", Evidence: json.RawMessage(`{}`)},
+		CSR:            csrPEM,
+		WorkloadClaims: "!!!not-base64!!!",
+	})
+	if err != nil {
+		t.Fatalf("marshal: %v", err)
+	}
+	req := httptest.NewRequest(http.MethodPost, "/attest", bytes.NewReader(body))
+	w := httptest.NewRecorder()
+	h.HandleAttest(w, req)
+	if w.Code != http.StatusBadRequest {
+		t.Fatalf("status: got %d, want 400; body=%s", w.Code, w.Body.String())
+	}
+}
+
+// A CSR whose RA-TLS attestation extension does not parse must be rejected at
+// issuance when workload claims are presented.
+func TestAttest_WorkloadClaims_RejectsGarbageRATLSExtension(t *testing.T) {
+	mock := newMockAttestationApi(t, "x")
+	h := newTestAttestHandler(t, mock.URL, nil)
+	h.AllowlistStore = floorStore(wlDigestA)
+
+	csrPEM, _ := generateCSRWithRATLSExtension(t, []byte("garbage-extension"))
+	digests := []string{wlDigestA}
+	w := postAttestClaimsWithCSR(t, h, issueChallenge(t, h), csrPEM, claimsDERFor(t, nil, digests), nil, digests)
+	if w.Code != http.StatusBadRequest {
+		t.Fatalf("status: got %d, want 400; body=%s", w.Code, w.Body.String())
+	}
+}
+
+func generateCSRWithRATLSExtension(t *testing.T, extValue []byte) (string, *ecdsa.PrivateKey) {
+	t.Helper()
+	key, err := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
+	if err != nil {
+		t.Fatalf("gen key: %v", err)
+	}
+	tmpl := &x509.CertificateRequest{
+		Subject:         pkix.Name{CommonName: "test-node"},
+		ExtraExtensions: []pkix.Extension{{Id: ratls.OIDRATLSAttestation, Value: extValue}},
+	}
+	der, err := x509.CreateCertificateRequest(rand.Reader, tmpl, key)
+	if err != nil {
+		t.Fatalf("create csr: %v", err)
+	}
+	return string(pem.EncodeToMemory(&pem.Block{Type: "CERTIFICATE REQUEST", Bytes: der})), key
+}
+
+// errStore fails every allowlist lookup, forcing the fail-closed branch.
+type errStore struct{}
+
+func (errStore) Contains(types.Digest) (bool, error) {
+	return false, errors.New("store unavailable")
+}
+
+func (errStore) LoadAll() (*pkgallowlist.Allowlist, string, error) {
+	return nil, "", errors.New("store unavailable")
+}
+
+func TestVerifyWorkloadClaims_FailsClosedOnStoreError(t *testing.T) {
+	h := AttestHandler{AllowlistStore: errStore{}}
+	digests := []string{wlDigestA}
+	err := h.verifyWorkloadClaims(claimsDERFor(t, nil, digests), nil, digests)
+	if err == nil {
+		t.Fatal("expected error when the allowlist store fails")
+	}
 }
