+++ internal/cmds/cds/signcsr_test.go
@@ -2,6 +2,7 @@ package cds
 
 import (
 	"bytes"
+	"context"
 	"crypto/ecdsa"
 	"crypto/elliptic"
 	"crypto/rand"
@@ -12,6 +13,7 @@ import (
 	"encoding/json"
 	"encoding/pem"
 	"log/slog"
+	"net"
 	"net/http"
 	"net/http/httptest"
 	"regexp"
@@ -25,6 +27,7 @@ import (
 	"github.com/confidential-dot-ai/c8s/internal/issuer"
 	"github.com/confidential-dot-ai/c8s/pkg/certutil"
 	"github.com/confidential-dot-ai/c8s/pkg/issuerapi"
+	"github.com/confidential-dot-ai/c8s/pkg/types"
 )
 
 type staticKeyProvider struct{ pub *ecdsa.PublicKey }
@@ -414,5 +417,207 @@ func TestSignCSR_CapsTTL(t *testing.T) {
 	got := leaf.NotAfter.Sub(leaf.NotBefore)
 	if got > h.MaxTTL+time.Minute {
 		t.Errorf("TTL not capped: got %v, want <= %v", got, h.MaxTTL)
+	}
+}
+
+func TestSignCSR_MissingEARField(t *testing.T) {
+	h, _, _ := newTestSignCSRHandler(t)
+	csr, _ := csrFor(t, pkix.Name{CommonName: "test-node"}, nil)
+	w := postSignCSR(t, h, "", csr.Raw, time.Hour)
+	if w.Code != http.StatusBadRequest {
+		t.Fatalf("status: got %d, want 400; body=%s", w.Code, w.Body.String())
+	}
+	if !bytes.Contains(w.Body.Bytes(), []byte("missing 'ear'")) {
+		t.Errorf("body should mention missing ear; got %s", w.Body.String())
+	}
+}
+
+func TestSignCSR_MissingCSRField(t *testing.T) {
+	h, earKey, _ := newTestSignCSRHandler(t)
+	earTok := signEAR(t, earKey, earClaimsLite{
+		Issuer:   "cds",
+		IssuedAt: time.Now().Unix(),
+		Expiry:   time.Now().Add(time.Minute).Unix(),
+	})
+	// Body with an EAR but a nil CSR DER.
+	body := []byte(`{"ear":"` + earTok + `","csr":"","ttl":"1h"}`)
+	req := httptest.NewRequest(http.MethodPost, "/sign-csr", bytes.NewReader(body))
+	w := httptest.NewRecorder()
+	h.HandleSignCSR(w, req)
+	if w.Code != http.StatusBadRequest {
+		t.Fatalf("status: got %d, want 400; body=%s", w.Code, w.Body.String())
+	}
+	if !bytes.Contains(w.Body.Bytes(), []byte("missing 'csr'")) {
+		t.Errorf("body should mention missing csr; got %s", w.Body.String())
+	}
+}
+
+func TestSignCSR_InvalidTTLMessage(t *testing.T) {
+	h, _, _ := newTestSignCSRHandler(t)
+	body := []byte(`{"ear":"x","csr":"x","ttl":12345}`)
+	req := httptest.NewRequest(http.MethodPost, "/sign-csr", bytes.NewReader(body))
+	w := httptest.NewRecorder()
+	h.HandleSignCSR(w, req)
+	if w.Code != http.StatusBadRequest {
+		t.Fatalf("status: got %d, want 400; body=%s", w.Code, w.Body.String())
+	}
+}
+
+func TestSignCSR_RejectsNonPEMCSRField(t *testing.T) {
+	h, _, _ := newTestSignCSRHandler(t)
+	body := []byte(`{"ear":"x","csr":"bm90LXBlbQ==","ttl":"1h"}`)
+	req := httptest.NewRequest(http.MethodPost, "/sign-csr", bytes.NewReader(body))
+	w := httptest.NewRecorder()
+	h.HandleSignCSR(w, req)
+	if w.Code != http.StatusBadRequest {
+		t.Fatalf("status: got %d, want 400; body=%s", w.Code, w.Body.String())
+	}
+}
+
+func TestSignCSR_RejectsPEMWithGarbageDER(t *testing.T) {
+	h, earKey, _ := newTestSignCSRHandler(t)
+	ear := signEAR(t, earKey, earClaimsLite{
+		Issuer:   "cds",
+		IssuedAt: time.Now().Unix(),
+		Expiry:   time.Now().Add(time.Minute).Unix(),
+	})
+	w := postSignCSR(t, h, ear, []byte("not-a-real-csr-der"), time.Hour)
+	if w.Code != http.StatusBadRequest {
+		t.Fatalf("status: got %d, want 400; body=%s", w.Code, w.Body.String())
+	}
+}
+
+func TestSignCSR_RejectsTamperedCSRSignature(t *testing.T) {
+	h, earKey, _ := newTestSignCSRHandler(t)
+	csr, _ := csrFor(t, pkix.Name{CommonName: "test-node"}, nil)
+	tampered := make([]byte, len(csr.Raw))
+	copy(tampered, csr.Raw)
+	tampered[len(tampered)-1] ^= 0xFF
+
+	ear := signEAR(t, earKey, earClaimsLite{
+		Issuer:   "cds",
+		IssuedAt: time.Now().Unix(),
+		Expiry:   time.Now().Add(time.Minute).Unix(),
+	})
+	w := postSignCSR(t, h, ear, tampered, time.Hour)
+	if w.Code != http.StatusBadRequest {
+		t.Fatalf("status: got %d, want 400; body=%s", w.Code, w.Body.String())
+	}
+}
+
+// csrForIP builds a CSR carrying a single IP SAN, exercising the source-IP
+// binding branch of issuer.ValidateCSR.
+func csrForIP(t *testing.T, subject pkix.Name, ip net.IP) (*x509.CertificateRequest, *ecdsa.PrivateKey) {
+	t.Helper()
+	key, err := ecdsa.GenerateKey(elliptic.P256(), rand.Reader)
+	if err != nil {
+		t.Fatalf("gen csr key: %v", err)
+	}
+	der, err := x509.CreateCertificateRequest(rand.Reader, &x509.CertificateRequest{
+		Subject:     subject,
+		IPAddresses: []net.IP{ip},
+	}, key)
+	if err != nil {
+		t.Fatalf("create csr: %v", err)
+	}
+	csr, err := x509.ParseCertificateRequest(der)
+	if err != nil {
+		t.Fatalf("parse csr: %v", err)
+	}
+	return csr, key
+}
+
+func TestSignCSR_SANValidationBindsSourceIP(t *testing.T) {
+	// postSignCSR uses httptest.NewRequest, whose RemoteAddr defaults to
+	// "192.0.2.1:1234", so that is the source IP the handler binds against.
+	t.Run("matching IP SAN accepted", func(t *testing.T) {
+		h, earKey, _ := newTestSignCSRHandler(t)
+		h.SANValidation = true
+		csr, csrKey := csrForIP(t, pkix.Name{CommonName: "test-node"}, net.ParseIP("192.0.2.1"))
+		ear := signEAR(t, earKey, earClaimsLite{
+			Issuer:    "cds",
+			IssuedAt:  time.Now().Unix(),
+			Expiry:    time.Now().Add(time.Minute).Unix(),
+			TEEPubKey: teePubKeyB64(t, csrKey),
+		})
+
+		w := postSignCSR(t, h, ear, csr.Raw, time.Hour)
+		if w.Code != http.StatusOK {
+			t.Fatalf("status: got %d, want 200; body=%s", w.Code, w.Body.String())
+		}
+	})
+
+	t.Run("mismatched IP SAN rejected", func(t *testing.T) {
+		h, earKey, _ := newTestSignCSRHandler(t)
+		h.SANValidation = true
+		csr, csrKey := csrForIP(t, pkix.Name{CommonName: "test-node"}, net.ParseIP("10.0.0.99"))
+		ear := signEAR(t, earKey, earClaimsLite{
+			Issuer:    "cds",
+			IssuedAt:  time.Now().Unix(),
+			Expiry:    time.Now().Add(time.Minute).Unix(),
+			TEEPubKey: teePubKeyB64(t, csrKey),
+		})
+
+		w := postSignCSR(t, h, ear, csr.Raw, time.Hour)
+		if w.Code != http.StatusForbidden {
+			t.Fatalf("status: got %d, want 403; body=%s", w.Code, w.Body.String())
+		}
+		if !strings.Contains(w.Body.String(), types.ErrorCodeCSRDenied) {
+			t.Errorf("body should mention %s; got %s", types.ErrorCodeCSRDenied, w.Body.String())
+		}
+	})
+}
+
+// cancelKeyProvider cancels the request context when the EAR key is fetched,
+// so the handler's post-validation ctx.Err() check fires.
+type cancelKeyProvider struct {
+	pub    *ecdsa.PublicKey
+	cancel context.CancelFunc
+}
+
+func (p cancelKeyProvider) PublicKey(string) (*ecdsa.PublicKey, error) {
+	p.cancel()
+	return p.pub, nil
+}
+
+func TestSignCSR_TimeoutBeforeSigningReturns504(t *testing.T) {
+	h, earKey, _ := newTestSignCSRHandler(t)
+	ctx, cancel := context.WithCancel(context.Background())
+	h.KeyProvider = cancelKeyProvider{pub: &earKey.PublicKey, cancel: cancel}
+
+	csr, csrKey := csrFor(t, pkix.Name{CommonName: "test-node"}, nil)
+	ear := signEAR(t, earKey, earClaimsLite{
+		Issuer:    "cds",
+		IssuedAt:  time.Now().Unix(),
+		Expiry:    time.Now().Add(time.Minute).Unix(),
+		TEEPubKey: teePubKeyB64(t, csrKey),
+	})
+	csrPEM := pem.EncodeToMemory(&pem.Block{Type: "CERTIFICATE REQUEST", Bytes: csr.Raw})
+	body, err := json.Marshal(map[string]any{"ear": ear, "csr": string(csrPEM), "ttl": "1h"})
+	if err != nil {
+		t.Fatalf("marshal: %v", err)
+	}
+	req := httptest.NewRequest(http.MethodPost, "/sign-csr", bytes.NewReader(body)).WithContext(ctx)
+	w := httptest.NewRecorder()
+	h.HandleSignCSR(w, req)
+	if w.Code != http.StatusGatewayTimeout {
+		t.Fatalf("status: got %d, want 504; body=%s", w.Code, w.Body.String())
+	}
+}
+
+func TestSignCSR_SignFailureReturns500(t *testing.T) {
+	h, earKey, _ := newTestSignCSRHandler(t)
+	h.CA = &issuer.CA{} // no cert/key loaded: SignCSR fails
+	csr, csrKey := csrFor(t, pkix.Name{CommonName: "test-node"}, nil)
+	ear := signEAR(t, earKey, earClaimsLite{
+		Issuer:    "cds",
+		IssuedAt:  time.Now().Unix(),
+		Expiry:    time.Now().Add(time.Minute).Unix(),
+		TEEPubKey: teePubKeyB64(t, csrKey),
+	})
+
+	w := postSignCSR(t, h, ear, csr.Raw, time.Hour)
+	if w.Code != http.StatusInternalServerError {
+		t.Fatalf("status: got %d, want 500; body=%s", w.Code, w.Body.String())
 	}
 }
