+++ internal/issuer/ratelimit_test.go
@@ -1,6 +1,7 @@
 package issuer
 
 import (
+	"context"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -100,5 +101,72 @@ func TestRateLimiterMaxEntries(t *testing.T) {
 	defer rl.mu.Unlock()
 	if len(rl.limiters) != 3 {
 		t.Errorf("expected 3 entries after cap, got %d", len(rl.limiters))
+	}
+}
+
+func TestRateLimitMiddlewareAllowsThenRejects(t *testing.T) {
+	// burst=1, rate=0 -> first request allowed, second rejected.
+	rl, err := NewIPRateLimiter(rate.Limit(0), 1, 10)
+	if err != nil {
+		t.Fatalf("NewIPRateLimiter: %v", err)
+	}
+	var served int
+	h := RateLimitMiddleware(rl, http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
+		served++
+		w.WriteHeader(http.StatusOK)
+	}))
+
+	first := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodGet, "/", nil)
+	req.RemoteAddr = "10.1.2.3:5555"
+	h.ServeHTTP(first, req)
+	if first.Code != http.StatusOK {
+		t.Fatalf("first request: code = %d, want 200", first.Code)
+	}
+
+	second := httptest.NewRecorder()
+	h.ServeHTTP(second, req)
+	if second.Code != http.StatusTooManyRequests {
+		t.Fatalf("second request: code = %d, want 429", second.Code)
+	}
+	if served != 1 {
+		t.Fatalf("handler served %d times, want 1", served)
+	}
+}
+
+func TestRateLimitMiddlewareHandlesPortlessRemoteAddr(t *testing.T) {
+	rl, err := NewIPRateLimiter(rate.Limit(100), 10, 10)
+	if err != nil {
+		t.Fatalf("NewIPRateLimiter: %v", err)
+	}
+	h := RateLimitMiddleware(rl, http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
+		w.WriteHeader(http.StatusOK)
+	}))
+	rec := httptest.NewRecorder()
+	req := httptest.NewRequest(http.MethodGet, "/", nil)
+	req.RemoteAddr = "/run/cds.sock" // no host:port
+	h.ServeHTTP(rec, req)
+	if rec.Code != http.StatusOK {
+		t.Fatalf("code = %d, want 200", rec.Code)
+	}
+}
+
+func TestIPRateLimiterEvictionLoopStopsOnCancel(t *testing.T) {
+	rl, err := NewIPRateLimiter(rate.Limit(100), 10, 10)
+	if err != nil {
+		t.Fatalf("NewIPRateLimiter: %v", err)
+	}
+	ctx, cancel := context.WithCancel(context.Background())
+	done := make(chan struct{})
+	go func() {
+		rl.EvictionLoop(ctx, time.Millisecond, time.Millisecond)
+		close(done)
+	}()
+	time.Sleep(10 * time.Millisecond)
+	cancel()
+	select {
+	case <-done:
+	case <-time.After(time.Second):
+		t.Fatal("EvictionLoop did not return after cancel")
 	}
 }
